pyproject.toml
[tool.poetry]
name = "cyclonedx-editor-validator"
version = "0"
description = "Tool for creating, modifying and validating CycloneDX SBOMs."
authors = [
    "Aleg Vilinski <aleg.vilinski@festo.com>",
    "Christian Beck <christian.beck@festo.com>",
    "Moritz Marseu <moritz.marseu@festo.com>"
]
license = "GPL-3.0-only"
readme = "README.md"
packages = [{include = "cdxev"}]

[tool.poetry.urls]
Documentation = 'https://festo-se.github.io/cyclonedx-editor-validator/'
Repository = 'https://github.com/Festo-se/cyclonedx-editor-validator/'
Issues = 'https://github.com/Festo-se/cyclonedx-editor-validator/issues'
Changelog = 'https://github.com/Festo-se/cyclonedx-editor-validator/releases'

[tool.poetry.scripts]
cdx-ev = "cdxev.__main__:main"

[tool.poetry.dependencies]
python = "^3.9.0"
python-dateutil = "2.8.2"
jsonschema = {version = "4.17.3", extras = ["format"]}

[tool.poetry.group.dev.dependencies]
flake8 = "6"
black = "23.3.0"
pep8-naming = "0.13.3"
mypy = "1.3.0"
mypy-gitlab-code-quality = "0.0.15"
types-python-dateutil = "2.8.19.13"
types-jsonschema = "4.17.0.8"
pytest = "7.3.2"
coverage = "7.2.7"
toml = "0.10.2"
typing-extensions = "4.6.3"
bandit = "1.7.5"
isort = "5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.semantic_release]
version_variable = [
    "pyproject.toml:version"
]
branch = "master"
upload_to_repository = false
upload_to_release = false
build_command = "pip install poetry && poetry build"

[tool.mypy]
files = ["cdxev/*"]
exclude = ["tests/*"]
disallow_untyped_defs = true
no_error_summary = true

[tool.coverage.run]
source = ["cdxev"]

[tool.coverage.report]
omit = ["*__init__.py*"]
